+++ contracts/PriceFeedInterface.sol
@@ -12,12 +12,12 @@ interface PriceFeedInterface {
 
     // Gets the latest time-price pair at which a price was published. `publishTime` will be 0 and `price` should be
     // ignored if no prices have ever been published for this symbol.
-    function latestPrice(bytes32 symbol) external view returns (uint publishTime, int256 price);
+    function latestPrice(bytes32 symbol) external view returns (uint publishTime, int price);
 
     // Gets the time-price pair at which a price was published that is nearest to `time` without being greater than
     // `time`. `publishTime` will be 0 and `price` should be ignored if no prices have been published before `time`.
-    function priceAtTime(bytes32 symbol, uint time) external view returns (uint publishTime, int256 price);
+    function priceAtTime(bytes32 symbol, uint time) external view returns (uint publishTime, int price);
 
     // An event fired when a price is published.
-    event PriceUpdated(bytes32 indexed symbol, uint time, int256 price);
+    event PriceUpdated(bytes32 indexed symbol, uint time, int price);
 }
